Abstraction of routes and creation of ngo controller to
