+++ backend/src/routes.js
@@ -1,7 +1,15 @@
 const express = require('express');
-const crypto = require('crypto');
+
+const NgoController = require('./controllers/NgoController')
 const routes = express.Router();
-const connection = require('./database/connection');
+
+// route to list the ngos on the data bank
+routes.get('/ngos', NgoController.index); 
+
+// route to insert ngo on the data bank
+routes.post('/ngos', NgoController.create); 
+
+module.exports = routes;
 
 /**
  * ROUTES / RESOURCES
@@ -19,31 +27,3 @@ const connection = require('./database/connection');
  * ROUTE PARAMETERS - PARAMETERS TO IDENTIFY RESOURCES
  * REQUEST BODY - REQUISITION BODY, USED TO CREATE OR UPDATE RESOURCES
  */
-
- // route to list the ngos on the data bank
- routes.get('/ngos', async (request, response) => {
-    const ngos = await connection.table('ngos').select('*');
-
-    return response.json(ngos);  
-});
-
-// route to insert ngo on the data bank
-routes.post('/ngos', async (request, response) => {
-    
-    const { name, email, whatsapp, city, uf } = request.body;
-    const id = crypto.randomBytes(4).toString('HEX');
-
-    //connect db
-    await connection('ngos').insert({
-        id,
-        name,
-        email,
-        whatsapp,
-        city,
-        uf,
-    });
-
-    return response.json({ id });
-});
-
-module.exports = routes;
